requirements.txt
# Core dependencies
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.39.0
pytorch-lightning>=2.0.0
datasets>=2.14.0
evaluate>=0.4.0
accelerate>=0.25.0
peft>=0.6.0

# Image processing
Pillow>=10.0.0

# Utilities
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
wandb>=0.15.0
pyyaml>=6.0

# Optional: for better performance
flash-attn>=2.3.0